+++ Editor/BuildReportVisualizer/BuildReportListView.cs
@@ -182,6 +182,7 @@ namespace UnityEditor.AddressableAssets.BuildReportVisualizer
             if (!string.IsNullOrEmpty(filePath) && File.Exists(filePath) && Path.GetExtension(filePath).ToLower() == ".json")
             {
                 AddReportFromFile(filePath, m_ListView, true, true);
+                AddressableAnalytics.ReportUsageEvent(AddressableAnalytics.UsageEventType.BuildReportImportedManually);
             }
         }
 
@@ -243,6 +244,7 @@ namespace UnityEditor.AddressableAssets.BuildReportVisualizer
                 }
             }
 
+            AddressableAnalytics.ReportUsageEvent(AddressableAnalytics.UsageEventType.BuildReportImportedManually);
             listView.Rebuild();
         }