+++ Editor/Diagnostics/Profiler/AddressablesProfilerDetailsTreeView.cs
@@ -124,7 +124,9 @@ namespace UnityEditor.AddressableAssets.Diagnostics
         private void Bind(string columnName, Label lbl, int itemIndex)
         {
             ContentData data = m_TreeView.GetItemDataForIndex<ContentData>(itemIndex);
-            if (lbl == null || data == null)
+            if (lbl == null)
+                return;
+            if (data == null)
             {
                 lbl.text = "ERROR";
                 return;